requirements.txt
pydantic>=2.6
python-dotenv>=1.0

# Server
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.9

# Settings / config
pydantic-settings>=2.2
PyYAML>=6.0

# HTTP + WS
httpx>=0.27
websockets>=12.0

# VLM (screenshot capture)
mss>=9.0
Pillow>=10.0

# Optional providers (graceful fallback if not configured)
google-genai>=0.6
google-cloud-texttospeech>=2.16
openai>=1.12

# STT (local)
faster-whisper>=1.1
numpy>=1.26

# Optional: title wordcloud generation
janome>=0.5
